backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
pydantic==2.0.3
pydantic-settings==2.0.2
sqlalchemy==2.0.19
pymysql==1.1.0
cryptography==41.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
paho-mqtt==2.1.0
